+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ChatView/ChatFormats/anthropic.ts
@@ -1,6 +1,7 @@
 import _ from 'lodash';
-import {Choice} from '../types';
-import {hasStringProp} from './utils';
+import { TraceCallSchema } from '../../wfReactInterface/traceServerClientTypes';
+import {ChatCompletion, ChatRequest, Choice} from '../types';
+import {hasStringProp, isMessage} from './utils';
 
 export const isAnthropicContentBlock = (block: any): boolean => {
   if (!_.isPlainObject(block)) {
@@ -58,3 +59,56 @@ export const anthropicContentBlocksToChoices = (
   }
   return choices;
 };
+
+
+export const isTraceCallChatFormatAnthropic = (call: TraceCallSchema): boolean => {
+  if (!('messages' in call.inputs)) {
+    return false;
+  }
+  const {messages} = call.inputs;
+  if (!_.isArray(messages)) {
+    return false;
+  }
+  if (
+    hasStringProp(call.inputs, 'model') &&
+    call.inputs.model.toLowerCase().includes('mistral')
+  ) {
+    return false;
+  }
+  return messages.every(isMessage);
+};
+
+export const normalizeAnthropicChatCompletion = (
+  request: ChatRequest,
+  completion: any
+): ChatCompletion => {
+  return {
+    id: completion.id,
+    choices: anthropicContentBlocksToChoices(
+      completion.content,
+      completion.stop_reason
+    ),
+    created: 0, // Anthropic doesn't provide `created`
+    model: completion.model,
+    system_fingerprint: '', // Anthropic doesn't provide `system_fingerprint`
+    usage: {
+      prompt_tokens: completion.usage.input_tokens,
+      completion_tokens: completion.usage.output_tokens,
+      total_tokens:
+        completion.usage.input_tokens + completion.usage.output_tokens,
+    },
+  };
+}
+
+export const normalizeAnthropicChatRequest = (request: ChatRequest & { system: string }) => {
+  return {
+    ...request,
+    messages: [
+      {
+        role: 'system',
+        content: request.system,
+      },
+      ...request.messages,
+    ],
+  };
+}
