+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/ChatView/ChatFormats/gemini.ts
@@ -3,7 +3,7 @@ import {
   KeyedDictType,
   TraceCallSchema,
 } from '../../wfReactInterface/traceServerClientTypes';
-import {Choice} from '../types';
+import {ChatCompletion, ChatRequest, Choice} from '../types';
 import {hasStringProp, hasNumberProp} from './utils';
 
 export const isGeminiRequestFormat = (inputs: KeyedDictType): boolean => {
@@ -93,3 +93,38 @@ export const isTraceCallChatFormatGemini = (call: TraceCallSchema): boolean => {
     isGeminiRequestFormat(call.inputs) && isGeminiCompletionFormat(call.output)
   );
 };
+
+export const normalizeGeminiChatCompletion = (
+  request: ChatRequest,
+  completion: any
+): ChatCompletion => {
+    // We normalize to the OpenAI format as our standard representation
+    // but the Gemini format does not have a direct mapping for some fields.
+    // For now we leave empty placeholders for type checking purposes.
+    return {
+      id: '',
+      choices: geminiCandidatesToChoices(completion.candidates),
+      created: 0,
+      model: request.model,
+      system_fingerprint: '',
+      usage: {
+        prompt_tokens: completion.usage_metadata.prompt_token_count,
+        completion_tokens: completion.usage_metadata.candidates_token_count,
+        total_tokens: completion.usage_metadata.total_token_count,
+      },
+    };
+}
+
+export const normalizeGeminiChatRequest = (request: any): ChatRequest => {
+  const modelIn = request.self.model_name;
+  const model = modelIn.split('/').pop() ?? '';
+  return {
+    model,
+    messages: [
+      {
+        role: 'system',
+        content: request.contents,
+      },
+    ],
+  };
+}
